out-Heading enhancement
Moved out-Heading into the “colors” library (becoming more of a display
library…). Also cleaned it up a bit so it’s more dynamic to screen
width (the background extends the full width… whatever that width might
be) (except when the width is not reported - I’m looking at you
GeekTool - then it defaults to 100 spaces).