Removed call for Panic Info from gt-info
Not sure when (guessing it was El Capitan), system profiler no longer
likes “system_profiler SPLogsDataType” to find the kernel panics. Since
Macs never panic (yes, that was sarcastic - I put this code in here for
a reason) and it’s been mostly useless for ages, I’ve disabled the call
to this routine. Now it serves as an example for how to disable a
routine :).